Create README.md
This project represents a slightly visually modified version of Minesweeper, called Pecksweeper (named for my high school CS teacher George Peck). It has various difficulties that can be set from the difficulty bar, and the gameplay is the same as any Minesweeper game. Read this for instructions on how to play Minesweeper  : https://www.instructables.com/How-to-play-minesweeper/